Refactor planning module and remove unused classes
      Summary: This commit simplifies and improves the planning logic in the SemanticKernel project by using the common PlannerSkill interface and the generic Plan class. It also removes the unused FunctionFlowRunner and SkillPlan classes and their dependencies, and makes the SKContextPlanningExtensions class public. Additionally, it renames the PlannerSkillConfig class to PlannerConfig and adds a TryGetPlan method to the SKContextPlanningExtensions class. These changes are part of a larger effort to streamline the planning and orchestration functionality in the SemanticKernel project.
